Update file structure and input_fn with TF.data
-Reorganize code
-Update README.md file with additional instructions
-Delete utils and integrate functionality within task.py
-Update model to support TF.data API
